tests: Merge common tests together
