+++ testsuite/python/CMakeLists.txt
@@ -125,9 +125,7 @@ python_test(FILE dds-and-bh-gpu.py MAX_NUM_PROC 4 LABELS gpu)
 python_test(FILE electrostaticInteractions.py MAX_NUM_PROC 2)
 python_test(FILE engine_langevin.py MAX_NUM_PROC 4)
 python_test(FILE engine_lb.py MAX_NUM_PROC 2 LABELS gpu)
-python_test(FILE engine_lb_hybrid.py MAX_NUM_PROC 2 LABELS gpu)
-python_test(FILE engine_lb_hybrid.py MAX_NUM_PROC 1 LABELS gpu SUFFIX 1_core)
-python_test(FILE engine_lb_nsquare.py MAX_NUM_PROC 1 LABELS gpu)
+python_test(FILE engine_lb.py MAX_NUM_PROC 1 LABELS gpu SUFFIX n_square)
 python_test(FILE icc.py MAX_NUM_PROC 4)
 python_test(FILE mass-and-rinertia_per_particle.py MAX_NUM_PROC 2 LABELS long)
 python_test(FILE integrate.py MAX_NUM_PROC 4)
@@ -194,10 +192,7 @@ python_test(FILE dipolar_interface.py MAX_NUM_PROC 1 LABELS gpu)
 python_test(FILE dipolar_mpi_exceptions.py MAX_NUM_PROC 2)
 python_test(FILE lb.py MAX_NUM_PROC 2 LABELS gpu)
 python_test(FILE lb_stats.py MAX_NUM_PROC 2 LABELS gpu long)
-python_test(FILE lb_stats_hybrid.py MAX_NUM_PROC 2 LABELS gpu long)
-python_test(FILE lb_stats_hybrid.py MAX_NUM_PROC 1 LABELS gpu long SUFFIX
-            1_core)
-python_test(FILE lb_stats_nsquare.py MAX_NUM_PROC 1 LABELS gpu long)
+python_test(FILE lb_stats.py MAX_NUM_PROC 1 LABELS gpu long SUFFIX n_square)
 python_test(FILE lb_vtk.py MAX_NUM_PROC 2 LABELS gpu)
 python_test(FILE force_cap.py MAX_NUM_PROC 2)
 python_test(FILE dpd.py MAX_NUM_PROC 4)
@@ -257,10 +252,8 @@ python_test(FILE lb_shear.py MAX_NUM_PROC 2 LABELS gpu)
 python_test(FILE lb_thermostat.py MAX_NUM_PROC 2 LABELS gpu)
 python_test(FILE lb_buoyancy_force.py MAX_NUM_PROC 4 LABELS gpu)
 python_test(FILE lb_momentum_conservation.py MAX_NUM_PROC 4 LABELS gpu)
-python_test(FILE lb_momentum_conservation_hybrid.py MAX_NUM_PROC 4 LABELS gpu)
-python_test(FILE lb_momentum_conservation_hybrid.py MAX_NUM_PROC 1 LABELS gpu
-            SUFFIX 1_core)
-python_test(FILE lb_momentum_conservation_nsquare.py MAX_NUM_PROC 1 LABELS gpu)
+python_test(FILE lb_momentum_conservation.py MAX_NUM_PROC 1 LABELS gpu SUFFIX
+            n_square)
 python_test(FILE p3m_electrostatic_pressure.py MAX_NUM_PROC 2)
 python_test(FILE sigint.py DEPENDENCIES sigint_child.py MAX_NUM_PROC 1)
 python_test(FILE lb_density.py MAX_NUM_PROC 1)
