+++ testsuite/python/lb_momentum_conservation.py
@@ -89,8 +89,8 @@ class Momentum(object):
 
 
 @utx.skipIfMissingGPU()
-@utx.skipIfMissingFeatures(['EXTERNAL_FORCES'])
-class TestLBGPUMomentumConservation(Momentum, ut.TestCase):
+@utx.skipIfMissingFeatures(["EXTERNAL_FORCES"])
+class TestRegularLBGPU(Momentum, ut.TestCase):
 
     lb_class = espressomd.lb.LBFluidGPU
 
@@ -98,13 +98,80 @@ class TestLBGPUMomentumConservation(Momentum, ut.TestCase):
         self.system.cell_system.set_regular_decomposition()
 
 
-@utx.skipIfMissingFeatures(['EXTERNAL_FORCES'])
-class TestLBCPUMomentumConservation(Momentum, ut.TestCase):
+@utx.skipIfMissingFeatures(["EXTERNAL_FORCES"])
+class TestRegularLBCPU(Momentum, ut.TestCase):
 
     lb_class = espressomd.lb.LBFluid
 
     def set_cellsystem(self):
         self.system.cell_system.set_regular_decomposition()
+
+
+@utx.skipIfMissingGPU()
+@utx.skipIfMissingFeatures(["EXTERNAL_FORCES"])
+class TestNSquareLBGPU(Momentum, ut.TestCase):
+
+    lb_class = espressomd.lb.LBFluidGPU
+
+    def set_cellsystem(self):
+        self.system.cell_system.set_n_square()
+
+
+@ut.skipIf(Momentum.n_nodes > 1,
+           "LB with N-square only works on 1 MPI rank")
+@utx.skipIfMissingFeatures(["EXTERNAL_FORCES"])
+class TestNSquareLBCPU(Momentum, ut.TestCase):
+
+    lb_class = espressomd.lb.LBFluid
+
+    def set_cellsystem(self):
+        self.system.cell_system.set_n_square()
+
+
+@utx.skipIfMissingGPU()
+@utx.skipIfMissingFeatures(["EXTERNAL_FORCES"])
+class TestHybrid0LBGPU(Momentum, ut.TestCase):
+
+    lb_class = espressomd.lb.LBFluidGPU
+
+    def set_cellsystem(self):
+        self.system.cell_system.set_hybrid_decomposition(
+            n_square_types={0}, cutoff_regular=1)
+
+
+@utx.skipIfMissingGPU()
+@utx.skipIfMissingFeatures(["EXTERNAL_FORCES"])
+class TestHybrid1LBGPU(Momentum, ut.TestCase):
+
+    lb_class = espressomd.lb.LBFluidGPU
+
+    def set_cellsystem(self):
+        self.system.cell_system.set_hybrid_decomposition(
+            n_square_types={1}, cutoff_regular=1)
+
+
+@ut.skipIf(Momentum.n_nodes > 1,
+           "LB with N-square only works on 1 MPI rank")
+@utx.skipIfMissingFeatures(["EXTERNAL_FORCES"])
+class TestHybrid0LBCPU(Momentum, ut.TestCase):
+
+    lb_class = espressomd.lb.LBFluid
+
+    def set_cellsystem(self):
+        self.system.cell_system.set_hybrid_decomposition(
+            n_square_types={0}, cutoff_regular=1)
+
+
+@ut.skipIf(Momentum.n_nodes > 1,
+           "LB with N-square only works on 1 MPI rank")
+@utx.skipIfMissingFeatures(["EXTERNAL_FORCES"])
+class TestHybrid1LBCPU(Momentum, ut.TestCase):
+
+    lb_class = espressomd.lb.LBFluid
+
+    def set_cellsystem(self):
+        self.system.cell_system.set_hybrid_decomposition(
+            n_square_types={1}, cutoff_regular=1)
 
 
 if __name__ == "__main__":
